// ui_tailwind_shadecn/src/enums/iframEnums.tsx
export enum ComponentType {
    COMPONENTS = '250px',
    VIEW = '500px',
    NAV = "100%",
    BUTTONS = "200px",
    CARDS = "300px",
    // TOOLTIP = "300px",
    // INPUT = "300px",
    // HERO = "300px",
    // FOOTER = "300px",
    // LOADER = "300",
    // SWITCH = "300",
    // CHECKBOX = "300",
    // FORM = "300"
  }